Fix getThankYou imports after hook move
