+++ client/my-sites/checkout/composite-checkout/test/composite-checkout-thank-you.js
@@ -7,7 +7,7 @@
 /**
  * Internal dependencies
  */
-import { getThankYouPageUrl } from '../use-get-thank-you-url';
+import getThankYouPageUrl from '../hooks/use-get-thank-you-url/get-thank-you-page-url';
 import { isEnabled } from 'calypso/config';
 
 let mockGSuiteCountryIsValid = true;
